requirements.txt
awsebcli
boto3
botocore
celery
coreapi
coreschema
cryptography
dateparser
Django==2.2.2
django-cities
django-extensions==2.1.6
djangorestframework
djangorestframework-serializer-extensions
django-rest-passwordreset
django-rest-swagger
django-storages
django-user-accounts
drfdocs
drf-openapi
dry-rest-permissions
geocoder
geoip2
geopy
gunicorn
hashids
ipython
ipython-genutils
ipywidgets
jupyter
jupyter-client
jupyter-console
jupyter-core
maxminddb
notebook
pillow
psycopg2
pygeocoder
python-dateutil
pytz
requests
six
social-auth-app-django
social-auth-core
street-address
termcolor
Werkzeug
whitenoise
django-hashid-field
django_celery_results
inflection
django-cors-headers
